[Major Update] Finished randomisePatterns & refactored code
Completed adding the text and images display and finished off the randomisePatterns function.

Refactored code to make it "siloed" and easier to find and debug code if needed. Need to check if it still works on the physical board.

Cleaned up code and removed useless code that was never accessed. Organised files and created a naming scheme for files.